Use KeepAlive component in Sidebar instead of v-if and v-show on same component
